perfectionist commit
Alif is a perfectionist about numbers. He loves to arrange things in order and sticks to his “Golden rule” that every set of numbers must be in ascending order. Unfortunately, that is not always the case. Alif defines that when a smaller number comes after a larger number in the set then number of violations are required to fix the order. Given a set of integers, Alif needs to find out the total number of such violations. Input:

⮚ The first line contains n, the number of integers.
⮚ The second line contains n space separated integers a0 … an-1 

Output: The output is an integer indicating the total number of violations.
Bonus: Time Complexity should be less than or equals to O(nlogn)

 Example:
Sample Input:
5
4 5 6 7 1

Sample Output :
4

Explanation :
1 violates which requires 4 violation to be fixed as 1 4 5 6 7.

Sample Input:
5
5 4 3 2 1

Sample Output :
10

Explanation :
4 violates which requires 1 violation to be fixed as 4 5 3 2 1.
Then again 3 violates which requires 2 violation to be fixed as 3 4 5 2 1.
Then again 2 violates which requires 3 violation to be fixed as 2 3 4 5 1.
Then again 1 violates which requires 4 violation to be fixed as 1 2 3 4 5.
Total violations required = 1+2+3+4 =10